+++ web/src/components/aboutDataPopover.test.tsx
@@ -1,25 +1,26 @@
 import React from 'react'
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
-import AboutDataPopover from 'features/fba/components/AboutDataPopover'
+import AboutDataPopover from '@/components/AboutDataPopover'
+import { FBAAboutDataContent } from '@/features/fbaCalculator/components/FbaAboutDataContent'
+import { ASAAboutDataContent } from '@/features/fba/components/ASAAboutDataContent'
 
 const ADVISORY_THRESHOLD = 20
 
 describe('AboutDataPopover', () => {
-  it('should render the About Data Popover', () => {
-    const { getByTestId } = render(<AboutDataPopover advisoryThreshold={ADVISORY_THRESHOLD}></AboutDataPopover>)
+  it('should render the About Data Popover with static content', () => {
+    const { getByTestId } = render(<AboutDataPopover content={FBAAboutDataContent}></AboutDataPopover>)
     const aboutData = getByTestId('about-data-popover')
     expect(aboutData).toBeInTheDocument()
   })
   it('should open the popover when clicked', () => {
-    render(<AboutDataPopover advisoryThreshold={ADVISORY_THRESHOLD}></AboutDataPopover>)
+    render(<AboutDataPopover content={FBAAboutDataContent}></AboutDataPopover>)
 
     fireEvent.click(screen.getByTestId('about-data-trigger'))
 
     expect(screen.getByTestId('about-data-content')).toBeVisible()
-    expect(screen.getByTestId('about-data-content')).toHaveTextContent(`${ADVISORY_THRESHOLD}%`)
   })
   it('should close the popover when clicking outside of it', async () => {
-    render(<AboutDataPopover advisoryThreshold={ADVISORY_THRESHOLD}></AboutDataPopover>)
+    render(<AboutDataPopover content={FBAAboutDataContent}></AboutDataPopover>)
     fireEvent.click(screen.getByTestId('about-data-trigger'))
     expect(screen.getByTestId('about-data-content')).toBeVisible()
 
@@ -28,8 +29,13 @@ describe('AboutDataPopover', () => {
       expect(screen.queryByTestId('popover-title')).not.toBeInTheDocument()
     })
   })
-  it('should contain the advisory threshold as a percent', () => {
-    render(<AboutDataPopover advisoryThreshold={ADVISORY_THRESHOLD}></AboutDataPopover>)
+  it('should contain the advisory threshold as a percent if passed as a prop', () => {
+    render(
+      <AboutDataPopover
+        content={ASAAboutDataContent}
+        props={{ advisoryThreshold: ADVISORY_THRESHOLD }}
+      ></AboutDataPopover>
+    )
 
     fireEvent.click(screen.getByTestId('about-data-trigger'))
